Billing, one cent precision
This patch allows the invoice item price and shop product price to be rounded
to the nearest cent instead of the nearest five cents previously.

Even if in Switzerland the smallest coin is 5 cents, invoices can still be
issued with a precision of one cent, which is useful for some products.

To ensure rounded prices the membership are still rounded to the nearest five
cents.

+++ app/models/checker/delivery_basket_content_avg_prices.rb
@@ -30,7 +30,7 @@ module Checker
     def diffs(before, after)
       before = before.values.sort
       after = after.values.sort
-      before.zip(after).map { |b, a| (b.to_f - a.to_f).abs }
+      before.zip(after).map { |b, a| (b.to_d - a.to_d).abs }
     end
   end
 end
